Implementación completa de logs para el proceso de autenticación

+++ src/main/java/uq/com/jdq/coresuite/infra/authenticationeventstype/AuthenticationEventsTypeRepository.java
@@ -2,6 +2,6 @@ package uq.com.jdq.coresuite.infra.authenticationeventstype;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthenticationEventsTypeRepository extends JpaRepository<AuthenticationEventsType, Long> {
+public interface AuthenticationEventsTypeRepository extends JpaRepository<AuthenticationEventsType, Integer> {
 
 }
